Fixed bug that resulted when we set up the admin to use the ItemEditForm (couldn't save anymore because admin doesn't pass user value in like our own views do -- this has now been compensated for)
Added a new field to Item to record the "worth" of an item.  This can be blank
if the user doesn't wish to set it.  Otherwise, it can be used to show the
difference between what the item is priced at and what it's really worth.

Added the new "worth" field to the item list web page.

Updated version