Removed all references to jquery.alerts (doesn't work with current jQuery). Created warning() which uses jquery-ui dialog instead.
Created a sales scenario to handle the problem when there is no tax rate (so
can't calculate price and commission automatically)

Throw up a warning dialog if there is no tax rate to calculate price and
commission automatically.

Modified form filling step (I set the {element} to {value}) to handle the
calendar popup, it was popping up in front of the button to press on Chrome, so
Chrome would fail.  Now, if date is being set, the step will make sure the popup
goes away before continuing.