Issue #9
Added a calendar icon next to the discount field on the Sales form which will get the actual sale
price from the user and automatically calculate the discount percentage.  It will also update the
price field with the actual sale price value.

Fixed a bug where the discount field is blank and the user goes to another field causing the commission
and price fields to automatically update via an Ajax call.  In this case, if the discount field is blank,
it's set to 0.